Update AudioInput and AudioOutput streams to use the Node 8.x streams API.
AudioInput now supports the standard destroy() function to close the stream in place of the previous abort().

AudioOutput no longer requires start() to be called, supports the standard end() function instead of quit() and supports the standard destroy() instead of abort().

Both implementations now ensure error events are emitted when no callback is provided to receive the error.

The native bindings were improved to ensure Pa_Terminate() is only called once per object (to avoid premature termination of the library) and to support the new stream APIs.